+++ tests/disjointTest.sql
@@ -1,8 +1,10 @@
+drop table tablea;
 create table tablea (
    id integer primary key,
    geom ast_polygon
 );
 
+drop table tableb;
 create table tableb (
    id integer primary key,
    geom ast_point