+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project is documented in this file.
 ## March 27, 2024
 ### New functionalities
 
+- `AST_LINE` is now mapped to a `multilinestring` geometry type in PostGIS. 
 - `AST_POLYGON` and `AST_PLANARSUBDIVISION` are now mapped to a `multipolygon` geometry type in PostGIS. 
 
 ## Jun 26, 2023